Add hip-tests to be built with TheRock (#2373)
## Motivation

This change enables hip-tests to be built on TheRock. This is the first
step in getting hip-tests building/running with TheRock.

## Technical Details

Enable hip-tests to be built with TheRock, when user enables testing via
flag `THEROCK_BUILD_TESTING`.


[hip-tests](https://github.com/ROCm/rocm-systems/tree/develop/projects/hip-tests/catch)
are collection of tests that are used to test the hip-runtime.

A corresponding PR is also required in hip-tests to make it compatible
with TheRock, which are being done
[here](https://github.com/ROCm/rocm-systems/pull/1624).

Subsequent PRs are required to run the built hip-tests which will follow
after this PR.

https://github.com/ROCm/rocm-systems/pull/2187 - PR will fix the paths
to be consistent across linux and windows.
 Below issues will be fixed with it
https://github.com/ROCm/TheRock/pull/2373/files#r2582804642  
https://github.com/ROCm/TheRock/pull/2373#issuecomment-3618267807 
https://github.com/ROCm/TheRock/pull/2001#discussion_r2578784486
https://github.com/ROCm/TheRock/pull/2001#discussion_r2569644056


## Test Plan

The build should complete end to end without failures, next set of
changes will enable the running of tests.

## Test Result

NA

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.