[CI] Added function gha_query_last_successful_workflow_run (#2610)
## Motivation
This is a common function that can be used across scripts to find the
last successful run of a specified workflow. This will allow devs to
fetch the entire workflow information and extract the data that is
relevant to them.

Split PR from: https://github.com/ROCm/TheRock/pull/2161/
<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->

## Technical Details
Uses GitHub API query parameters to pre-filter for successful runs and
limit results. Further filtering is done via a for loop to filter for
runs on the specified branch.
<!-- Explain the changes along with any relevant GitHub links. -->

## Test Plan
Incorporated unit tests to check for successful run found on main
branch, no matching branch, and non-existent workflow.
All unit tests pass.
<!-- Explain any relevant testing done to verify this PR. -->

## Test Result
Here is an example output when run on ci_nightly.yml workflow
```
{
  "id": 19451926065,
  "name": "CI Nightly",
  "head_branch": "main",
  "head_sha": "ddd714ee73a9d4d77fbbd1d0022a541f6c8e102b",
  "status": "completed",
  "conclusion": "success",
  "run_number": 173,
  "event": "schedule",
  "created_at": "2025-11-18T02:33:25Z",
  "updated_at": "2025-11-18T06:06:07Z",
  "html_url": "https://github.com/ROCm/TheRock/actions/runs/19451926065",
  "actor": {
    "login": "geomin12"
  },
  "head_commit": {
    "id": "ddd714ee73a9d4d77fbbd1d0022a541f6c8e102b",
    "message": "Introduce docs/development/style_guide.md (#2095)",
    "author": {
      "name": "Scott Todd"
    }
  },
  "repository": {
    "full_name": "ROCm/TheRock"
  },
  "jobs_url": "https://api.github.com/repos/ROCm/TheRock/actions/runs/19451926065/jobs",
  "artifacts_url": "https://api.github.com/repos/ROCm/TheRock/actions/runs/19451926065/artifacts"
}
```

<!-- Briefly summarize test outcomes. -->

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.